backend/services/web_search_service/requirements.txt
fastapi
uvicorn
tavily-python
